+++ kfusion/src/benchmark.cpp
@@ -102,8 +102,8 @@ int main(int argc, char ** argv) {
 
 #ifdef SYCL
   uint2 computationSize = make_uint2(
-			inputSize.x() / config.compute_size_ratio,
-			inputSize.y() / config.compute_size_ratio);
+			((uint)inputSize.x()) / config.compute_size_ratio,
+			((uint)inputSize.y()) / config.compute_size_ratio);
 #else
 	const uint2 computationSize = make_uint2(
 			inputSize.x / config.compute_size_ratio,
